fixed bugs with enable throttling var, added redis skips for when redis fails. added scripts to start/stop redis

+++ src/marqo/connections.py
@@ -26,6 +26,8 @@ class RedisDriver:
     def __init__(self):
         self.driver: redis.Redis = None
         self.lua_shas: dict = None
+        self.host = "localhost"
+        self.port = 6379
 
         # Specify any LUA scripts to be used by the driver here.
         self.scripts = [
@@ -41,42 +43,61 @@ class RedisDriver:
         ]  
 
     def init_from_app(self, host, port):
-
         t0 = time.time()
 
-        self.connect(host, port)
-        self.load_lua_scripts()
-        
-        # remove any info from last Marqo instance
-        self.driver.flushdb()
+        self.host = host
+        self.port = port
 
-        t1 = time.time()
-        logger.info(f"Took {((t1-t0)*1000):.3f}ms to connect to redis and load scripts.\n")
+        if self.connect():
+            self.load_lua_scripts()
+            # remove any info from last Marqo instance
+            self.driver.flushdb()
 
-    def connect(self, host, port) -> redis.Redis:
+            t1 = time.time()
+            logger.info(f"Took {((t1-t0)*1000):.3f}ms to connect to redis and load scripts.")
+
+    def connect(self) -> redis.Redis:
+        # Does not raise an exception even when no redis-server is available. Monitor this in case that changes.
         self.driver = redis.Redis(
-            host=host,
-            port=port,
+            host=self.host,
+            port=self.port,
         )
-        return self.driver
-    
+        
+        if self.is_alive():
+            return self.driver
+        else:
+            # When other functions see driver = None, they can quit out.
+            self.driver = None
+            return None
+        
     def load_lua_scripts(self) -> dict:
+        print("Loading lua scripts now.")
         self.lua_shas = dict()
         for script in self.scripts:
             script_file = open(script["path"], "r")
             script_text = script_file.read()
 
             self.lua_shas[script["name"]] = self.driver.script_load(script_text)
+            print(f"Just loaded in {script['name']}")
             script_file.close()
         return self.lua_shas
 
     def get_db(self):
-        if not self.driver:
-            return self.connect()
+        # Try to connect if no good connection exists yet.
+        if not self.driver or not self.is_alive():
+            self.init_from_app(self.host, self.port)
         return self.driver
     
     def get_lua_shas(self):
         return self.lua_shas
+    
+    def is_alive(self):
+        try: 
+            self.driver.ping()
+            return True
+        except Exception as e:
+            logger.warn(f"Could not confirm connection to redis. Ensure you have a redis server running for throttling. Reason: {e}")
+            return False
 
 # Starts up redis driver
 # Can be turned off with MARQO_ENABLE_THROTTLING = 'FALSE'
